Messages to app work
Request messages sent to app from browser are received via WebSocket,
added to Ciera external message queue, and they cause the model to
transition and produce a reply message.  The reply message is, I think,
sent out via WebSockt to the browser, but presently, the browser app
does not display it.  I think the issue is in index.html and/or app.js.