Does not work
This commit produces no errors, but the reply message is never displayed
on the client side, probably because it is never even sent to the
message broker.
The @SendTo annotation must be used in conjunction with a mapping
annotation like @MessageMapping on the same method.  This does not
accomplish our goal, so we will use the SimpMessagingTemplate construct
instead.