Demonstrates comms with two specific clients
This is a primitive demonstration that the server can direct messages to
a specific client by sending to a particular topic to which the client
subscribes.  In this case, there are two web pages, north.html, and
south.html, each with an associated script (north.js and south.js).
Both scripts subscribe to /topic/replies as well as the
location-specific topic /topic/replies/north or /topic/replies/south.
Messages containing "south" are sent only to /topic/replies/south, while
messages containing "north" are sent only to /topic/replies/north, and
hence visible only to a specific client based on location. Messages
containing neither "north" nor "south" are sent to /topic/replies, and
hence visible to both clients.