Works, but...
This commit works as intended, but it is littered with hacks.
Invoke the server (bash ./run.sh) and then point one browser at
http://localhost:8080/Reply and another broswer at
http://localhost:8080/Request?msg=Hey.
Refresh the browser pointed at Reply after sending the Request message
with the other browser.
Notice that the Request page always shows the previous (not the most
recent) version of the request message because the Spring model data
gets updated before displaying the page.
In any case, with some modifications, this approach could be used for
the Carpark application employing either HTTP polling or long polling.